Made some big changes.
I first understood the assignment wrong,
so the result was that hovering would change the color
of a cell, and then the color would fade away in a 3 second
transition.

The goal was so that you can actually draw stuff on the board
without the color fading away - and now it does exactly that.